+++ plutarch-ledger-api/src/Plutarch/LedgerApi/V1/Scripts.hs
@@ -61,18 +61,19 @@ newtype PDatum (s :: S) = PDatum (Term s PData)
     )
   deriving anyclass
     ( -- | @since 2.0.0
-      PlutusType
-    , -- | @since 2.0.0
       PIsData
     , -- | @since 2.0.0
       PEq
     , -- | @since 2.0.0
       PShow
+    , -- | @since WIP
+      SOP.Generic
     )
-
--- | @since 2.0.0
-instance DerivePlutusType PDatum where
-  type DPTStrat _ = PlutusTypeNewtype
+  deriving
+    ( -- | @since 2.0.0
+      PlutusType
+    )
+    via (DeriveNewtypePlutusType PDatum)
 
 -- | @since 3.3.0
 deriving via
@@ -125,18 +126,19 @@ newtype PRedeemer (s :: S) = PRedeemer (Term s PData)
     )
   deriving anyclass
     ( -- | @since 2.0.0
-      PlutusType
-    , -- | @since 2.0.0
       PIsData
     , -- | @since 2.0.0
       PEq
     , -- | @since 2.0.0
       PShow
+    , -- | @since WIP
+      SOP.Generic
     )
-
--- | @since 2.0.0
-instance DerivePlutusType PRedeemer where
-  type DPTStrat _ = PlutusTypeNewtype
+  deriving
+    ( -- | @since 2.0.0
+      PlutusType
+    )
+    via (DeriveNewtypePlutusType PRedeemer)
 
 -- | @since 3.3.0
 deriving via